Refactor: Code Cleanup
