+++ components/SearchSandbox/components/RSWrapper.js
@@ -19,21 +19,19 @@ import {
  DataSearch, MultiList, ReactiveList, CategorySearch,
 } from '@appbaseio/reactivesearch';
 
-import getNestedValue from '../utils';
 import dataSearchTypes from '../utils/datasearch-types';
 import multiListTypes from '../utils/multilist-types';
 import reactiveListTypes from '../utils/reactivelist-types';
 import categorySearchTypes from '../utils/categorysearch-types';
-import { generateDataField, generateFieldWeights } from '../utils/dataField';
+import { generateDataField } from '../utils/dataField';
 import constants from '../utils/constants';
 import { getComponentCode } from '../template';
 import PreviewList from './PreviewList';
+import { SandboxContext } from '../index';
+import getComponentProps from '../utils/getComponentProps'
 
 import {
-	deleteStyles,
-	rowStyles,
-	formWrapper,
-	componentStyles,
+ deleteStyles, rowStyles, formWrapper, componentStyles,
 } from '../styles';
 
 const componentMap = {
@@ -50,7 +48,13 @@ const propsMap = {
 	CategorySearch: categorySearchTypes,
 };
 
-export default class RSWrapper extends Component {
+const RSWrapper = props => (
+	<SandboxContext.Consumer>
+		{contextValue => <RSComponentRender {...contextValue} {...props} />}
+	</SandboxContext.Consumer>
+);
+
+class RSComponentRender extends Component {
 	constructor(props) {
 		super(props);
 
@@ -112,7 +116,9 @@ export default class RSWrapper extends Component {
 
 	getCategoryField = () => {
 		const { component, mappings } = this.props;
-		const { categoryField: { types } } = propsMap[component];
+		const {
+			categoryField: { types },
+		} = propsMap[component];
 
 		const fields = Object.keys(mappings).filter((field) => {
 			let fieldsToCheck = [mappings[field]];
@@ -128,7 +134,7 @@ export default class RSWrapper extends Component {
 		});
 
 		return fields;
-	}
+	};
 
 	setError = (error) => {
 		this.setState(
@@ -187,7 +193,7 @@ export default class RSWrapper extends Component {
 	};
 
 	handleSwitchPropChange = (name, value) => {
-	const { componentProps } = this.state;
+		const { componentProps } = this.state;
 		this.setState({
 			componentProps: {
 				...componentProps,
@@ -211,7 +217,7 @@ export default class RSWrapper extends Component {
 		this.setState({
 			isInputActive: false,
 		});
-	}
+	};
 
 	handleDropdownChange = (e, name) => {
 		const { componentProps } = this.state;
@@ -226,12 +232,14 @@ export default class RSWrapper extends Component {
 	};
 
 	handleDropdownSearch = (e) => {
-		const { target: { name, value } } = e;
+		const {
+			target: { name, value },
+		} = e;
 		this.setState({
 			isInputActive: true,
 			[name]: value,
 		});
-	}
+	};
 
 	handleSearchDataFieldChange = (item) => {
 		const field = item.key;
@@ -250,13 +258,11 @@ export default class RSWrapper extends Component {
 	};
 
 	handleSearchDataFieldDelete = (deleteIndex) => {
-	const { componentProps } = this.state;
+		const { componentProps } = this.state;
 		this.setState({
 			componentProps: {
 				...componentProps,
-				dataField: componentProps.dataField.filter(
-					(i, index) => index !== deleteIndex,
-				),
+				dataField: componentProps.dataField.filter((i, index) => index !== deleteIndex),
 				fieldWeights: componentProps.fieldWeights.filter(
 					(i, index) => index !== deleteIndex,
 				),
@@ -279,7 +285,7 @@ export default class RSWrapper extends Component {
 	};
 
 	handleSearchWeightChange = (index, value) => {
-	const { componentProps } = this.state;
+		const { componentProps } = this.state;
 		const fieldWeights = Object.assign([], componentProps.fieldWeights, {
 			[index]: value,
 		});
@@ -292,7 +298,7 @@ export default class RSWrapper extends Component {
 	};
 
 	handleAddFieldRow = () => {
-	const { componentProps } = this.state;
+		const { componentProps } = this.state;
 		const field = this.getAvailableDataField().find(
 			item => !componentProps.dataField.includes(item),
 		);
@@ -328,7 +334,7 @@ export default class RSWrapper extends Component {
 	renderComponentCode = () => {
 		const {
 			component, id, mappings, componentProps, customProps,
-		} = this.props;
+		} = this.props; // prettier-ignore
 		const customComponentProps = customProps[component];
 
 		const config = {
@@ -439,11 +445,9 @@ export default class RSWrapper extends Component {
 	};
 
 	renderFormItem = (item, name) => {
-	const { componentProps, mappings } = this.props;
+		const { componentProps } = this.props;
 		let FormInput = null;
-		const value = componentProps[name] !== undefined
-				? componentProps[name]
-				: item.default;
+		const value = componentProps[name] !== undefined ? componentProps[name] : item.default;
 
 		switch (item.input) {
 			case 'bool': {
@@ -475,16 +479,26 @@ export default class RSWrapper extends Component {
 
 				const {
 					onPropChange, id, component, mappingsURL,
-				} = this.props;
-				const { componentProps: stateComponentProps, isInputActive, searchTerm } = this.state;
+				} = this.props; // prettier-ignore
+				const {
+					componentProps: stateComponentProps,
+					isInputActive,
+					searchTerm,
+				} = this.state;
 
 				if (name === 'categoryField') {
-					noOptionsMessage = <p style={{ lineHeight: '1.5' }}>There are no compatible fields present in your data mappings. <a href={mappingsURL}>You can edit your mappings</a> (agggregation components)</p>;
+					noOptionsMessage = (
+						<p style={{ lineHeight: '1.5' }}>
+							There are no compatible fields present in your data mappings.{' '}
+							<a href={mappingsURL}>You can edit your mappings</a> (agggregation
+							components)
+						</p>
+					);
 
 					this.getCategoryField().forEach(field => dropdownOptions.push({
-						label: field,
-						key: field,
-					}));
+							label: field,
+							key: field,
+						}));
 
 					if (dropdownOptions.length) {
 						if (!stateComponentProps.categoryField) {
@@ -493,24 +507,25 @@ export default class RSWrapper extends Component {
 							});
 						}
 
-						selectedDropdown = dropdownOptions
-							.find(option => option.key === stateComponentProps[name]);
+						selectedDropdown = dropdownOptions.find(
+							option => option.key === stateComponentProps[name],
+						);
 						selectedValue = selectedDropdown
-						? selectedDropdown.label
-						: dropdownOptions[0].label;
+							? selectedDropdown.label
+							: dropdownOptions[0].label;
 					}
 				} else {
 					dropdownOptions = propsMap[component][name].options;
-					selectedDropdown = dropdownOptions
-						.find(option => option.key === stateComponentProps[name]);
+					selectedDropdown = dropdownOptions.find(
+						option => option.key === stateComponentProps[name],
+					);
 					selectedValue = selectedDropdown
 						? selectedDropdown.label
 						: propsMap[component][name].default;
 				}
 
 				if (isInputActive) {
-					dropdownOptions = dropdownOptions
-						.filter(option => option.label.toLowerCase().includes(searchTerm.toLowerCase()));
+					dropdownOptions = dropdownOptions.filter(option => option.label.toLowerCase().includes(searchTerm.toLowerCase()));
 				}
 
 				if (!dropdownOptions.length) {
@@ -543,7 +558,9 @@ export default class RSWrapper extends Component {
 							onChange={this.handleDropdownSearch}
 						/>
 					</Dropdown>
-				) : noOptionsMessage;
+				) : (
+					noOptionsMessage
+				);
 				break;
 			}
 
@@ -627,86 +644,35 @@ export default class RSWrapper extends Component {
 
 	render() {
 		const {
-			componentProps, component, id, mappings, customProps, full, showDelete, onDelete, showCodePreview, showCustomList
+			componentProps,
+			component,
+			id,
+			mappings,
+			customProps,
+			full,
+			showDelete,
+			onDelete,
+			showCodePreview,
+			showCustomList,
+			setRenderKey,
 		} = this.props;
 		const { showModal, componentProps: stateComponentProps, previewModal } = this.state;
 		if (!componentProps.dataField) return null;
+
 		const RSComponent = componentMap[component];
 		let tutorialClass = '';
 		let editTutorialClass = '';
-		let otherProps = {};
+
 		if (id === 'search') {
 			tutorialClass = 'search-tutorial-1';
 			editTutorialClass = 'search-tutorial-2';
-			otherProps = {
-				fieldWeights: generateFieldWeights(
-					componentProps.dataField,
-					componentProps.fieldWeights,
-					mappings,
-				),
-				highlightField: componentProps.dataField,
-			};
 		}
 		if (id === 'result') {
 			tutorialClass = 'search-tutorial-4';
 			editTutorialClass = 'search-tutorial-5';
 		}
-		if (component === 'CategorySearch') {
-			otherProps = {
-				categoryField: generateDataField(
-					'MultiList',
-					componentProps.categoryField,
-					mappings,
-				),
-				fieldWeights: generateFieldWeights(
-					componentProps.dataField,
-					componentProps.fieldWeights,
-					mappings,
-				),
-				highlightField: componentProps.dataField,
-			};
-		}
-		const { componentProps: { metaFields, ...restProps } } = this.props;
-		const isMetaDataPresent = metaFields && metaFields.title && metaFields.description;
 
-		if (id === 'result' && isMetaDataPresent) {
-			const {
-				url: urlKey, title: titleKey, image: imageKey, description: descriptionKey,
-			} = metaFields;
-			otherProps = {
-					onData: (res) => {
-						const url = getNestedValue(res, urlKey);
-						const title = getNestedValue(res, titleKey);
-						const description = getNestedValue(res, descriptionKey);
-						const image = getNestedValue(res, imageKey);
-						const { renderJSONEditor, renderDeleteJSON } = this.props;
-
-						return (
-						<Row type="flex" key={res._id} style={{ margin: '20px auto', borderBottom: '1px solid #ededed' }}>
-							<Col span={image ? 6 : 0}>
-								<img style={{ width: '100%' }} src={image} alt={title} />
-							</Col>
-							<Col span={image ? 18 : 24}>
-								<h3 style={{ fontWeight: '600' }} dangerouslySetInnerHTML={{__html: title}} />
-								<p style={{ fontSize: '1em' }} dangerouslySetInnerHTML={{ __html: description }} />
-							</Col>
-							<div style={{ width: '100%', marginBottom: '10px', textAlign: 'right' }}>
-								{url ? <Button shape="circle" icon="link" style={{ marginRight: '5px' }} onClick={() => window.open(url, '_blank')} />
-								: null}
-								{renderJSONEditor(res)}
-								{renderDeleteJSON(res)}
-							</div>
-						</Row>
-					);
-				},
-			};
-		}
-
-		if (id === 'result' && componentProps.sortBy === 'best') {
-			delete restProps.sortBy;
-		}
-
-		const showPreview =	component === 'ReactiveList';
+		const showPreview = component === 'ReactiveList';
 		const customComponentProps = customProps[component];
 
 		return (
@@ -746,16 +712,10 @@ export default class RSWrapper extends Component {
 					<Col span={full ? 24 : 20} id={id} className={tutorialClass}>
 						<RSComponent
 							componentId={id}
-							{...restProps}
-							dataField={generateDataField(
-								component,
-								componentProps.dataField,
-								mappings,
-							)}
 							className={componentStyles}
-							fuzziness={componentProps.fuzziness || 0}
-							size={parseInt(componentProps.size || 10, 10)}
-							{...otherProps}
+							{...getComponentProps({
+								component, componentProps, mappings, setRenderKey,
+							})}
 							{...customComponentProps}
 						/>
 					</Col>
@@ -789,15 +749,10 @@ export default class RSWrapper extends Component {
 						options={Object.keys(mappings)}
 						componentProps={stateComponentProps}
 						componentId={id}
-						getNestedValue={getNestedValue}
 						handlePreviewModal={this.handlePreviewModal}
 						handleSavePreview={this.handleSavePreview}
 						visible={previewModal}
-						dataField={generateDataField(
-							component,
-							componentProps.dataField,
-							mappings,
-						)}
+						dataField={generateDataField(component, componentProps.dataField, mappings)}
 					/>
 				) : null}
 			</div>
@@ -808,3 +763,5 @@ export default class RSWrapper extends Component {
 RSWrapper.defaultProps = {
 	showDelete: true,
 };
+
+export default RSWrapper;
