+++ components/SearchSandbox/containers/Editor.js
@@ -11,25 +11,14 @@ import {
 	Dropdown,
 	Icon,
 	Menu,
-	Tree,
-	Popover,
-	Tooltip,
-	notification,
-	Popconfirm,
 	message,
 } from 'antd';
 import { ReactiveBase, SelectedFilters } from '@appbaseio/reactivesearch';
-import ExpandCollapse from 'react-expand-collapse';
 import PropTypes from 'prop-types';
 
-import Appbase from 'appbase-js';
-
-import Ace from './AceEditor';
 import multiListTypes from '../utils/multilist-types';
 import RSWrapper from '../components/RSWrapper';
-import { listItem, formWrapper } from '../styles';
-
-const { TreeNode } = Tree;
+import { formWrapper } from '../styles';
 
 export default class Editor extends Component {
 	constructor(props) {
@@ -41,19 +30,9 @@ export default class Editor extends Component {
 			listComponentProps: {
 				dataField: dataFields.length ? dataFields[0] : '',
 			},
-			editorValue: '',
-			isValidJSON: true,
-			editorObjectId: '',
 			renderKey: Date.now(),
 			showVideo: false,
-			isEditable: false,
 		};
-		const { appName, url, credentials } = props;
-		this.appbaseRef = Appbase({
-			app: appName,
-			url,
-			credentials,
-		});
 	}
 
 	getAvailableDataField = () => {
@@ -74,26 +53,6 @@ export default class Editor extends Component {
 		});
 
 		return fields;
-	};
-
-	copyJSON = (code) => {
-		const el = document.createElement('textarea');
-		el.value = JSON.stringify(code);
-		document.body.appendChild(el);
-		el.select();
-		document.execCommand('copy');
-		document.body.removeChild(el);
-		this.setState(
-			{
-				copied: true,
-			},
-			() => setTimeout(
-					() => this.setState({
-							copied: false,
-						}),
-					300,
-				),
-		);
 	};
 
 	showModal = () => {
@@ -180,106 +139,6 @@ export default class Editor extends Component {
 				...listComponentProps,
 				[name]: type === 'number' ? parseInt(value, 10) : value,
 			},
-		});
-	};
-
-	handleUpdateJSON = (updatedJSONString) => {
-		const updatedJSON = JSON.parse(updatedJSONString);
-		const { mappingsType } = this.props;
-		const { editorObjectId } = this.state;
-		let responseMessage = {
-			message: 'Edit successfully saved',
-			description: 'The desired result data was successfully updated.',
-			duration: 4,
-		};
-		this.appbaseRef
-			.update({
-				type: mappingsType,
-				id: editorObjectId,
-				body: {
-					doc: updatedJSON,
-				},
-			})
-			.then((res) => {
-				this.setState({
-					isEditable: false,
-					renderKey: res._timestamp, // eslint-disable-line
-				});
-				notification.open(responseMessage);
-			})
-			.catch(() => {
-				responseMessage = {
-					message: 'Update JSON',
-					description: 'There were error in Updating JSON. Try again Later.',
-					duration: 2,
-				};
-				notification.open(responseMessage);
-			});
-	};
-
-	handleDeleteJSON = (id) => {
-		let responseMessage = {
-			message: 'Delete JSON',
-			description: 'You have successfully deleted JSON.',
-			duration: 4,
-		};
-		const { mappingsType } = this.props;
-		this.appbaseRef
-			.delete({
-				type: mappingsType,
-				id,
-			})
-			.then((res) => {
-				this.setState({
-					renderKey: res._timestamp,
-				});
-				notification.open(responseMessage);
-			})
-			.catch(() => {
-				responseMessage = {
-					message: 'Delete JSON',
-					description: 'There were error in Deleting JSON. Try again Later.',
-					duration: 2,
-				};
-				notification.open(responseMessage);
-			});
-	};
-
-	handleEditing = () => {
-		this.setState(({ isEditable }) => ({
-			isEditable: !isEditable,
-		}));
-	};
-
-	handleEditingJSON = (value) => {
-		let isValidJSON = true;
-		try {
-			JSON.parse(value);
-		} catch (e) {
-			isValidJSON = false;
-		}
-		this.setState({
-			editorValue: value,
-			isValidJSON,
-		});
-	};
-
-	handleInitialEditorValue = (res) => {
-		const { _id: id, _index: del, ...objectJSON } = res;
-
-		const stringifiedJSON = JSON.stringify(objectJSON, null, '\t');
-
-		this.setState({
-			editorObjectId: id,
-			editorValue: stringifiedJSON,
-		});
-	};
-
-	resetEditorValues = () => {
-		this.setState({
-			editorObjectId: '',
-			editorValue: '',
-			isEditable: false,
 		});
 	};
 
@@ -383,128 +242,11 @@ export default class Editor extends Component {
 		);
 	};
 
-	renderAsTree = (res, key = '0') => {
-		if (!res) return null;
-		const iterable = Array.isArray(res) ? res : Object.keys(res);
-		return iterable.map((item, index) => {
-			const type = typeof res[item];
-			if (type === 'string' || type === 'number') {
-				const title = (
-					<div>
-						<span>{item}:</span>
-						&nbsp;
-						<span dangerouslySetInnerHTML={{ __html: res[item] }} />
-					</div>
-				);
-				return <TreeNode title={title} key={`${key}-${index + 1}`} />;
-			}
-			const hasObject = res[item] === undefined && typeof item !== 'string';
-			const node = hasObject ? item : res[item];
-			return (
-				<TreeNode
-					title={
-						typeof item !== 'string'
-							? 'Object'
-							: `${node || Array.isArray(res) ? item : `${item}: null`}`
-					}
-					key={`${key}-${index + 1}`}
-				>
-					{this.renderAsTree(node, `${key}-${index + 1}`)}
-				</TreeNode>
-			);
+	setRenderKey = (newKey) => {
+		this.setState({
+			renderKey: newKey,
 		});
-	};
-
-	renderDeleteJSON = res => (
-		<Popconfirm
-			title="Are you sure you want to delete this JSON?"
-			placement="bottomRight"
-			onConfirm={() => this.handleDeleteJSON(res._id)}
-			okText="Yes"
-		>
-			<Button shape="circle" icon="delete" style={{ marginRight: '5px' }} />
-		</Popconfirm>
-	);
-
-	renderJSONEditor = (res) => {
-		const {
-			isEditable, copied, isValidJSON, editorValue,
-		} = this.state;
-		return (
-			<Popover
-				placement="leftTop"
-				trigger="click"
-				onVisibleChange={visible => (
-					visible ? this.handleInitialEditorValue(res) : this.resetEditorValues())
-				}
-				content={
-					isEditable ? (
-						<Ace
-							mode="json"
-							value={editorValue}
-							onChange={value => this.handleEditingJSON(value)}
-							theme="monokai"
-							name="editor-JSON"
-							fontSize={14}
-							showPrintMargin
-							style={{ maxHeight: '250px' }}
-							showGutter
-							highlightActiveLine
-							setOptions={{
-								showLineNumbers: true,
-								tabSize: 2,
-							}}
-							editorProps={{ $blockScrolling: true }}
-						/>
-					) : (
-						<pre style={{ width: 300 }}>{JSON.stringify(res, null, 4)}</pre>
-					)
-				}
-				title={(
-					<Row>
-						<Col span={isEditable ? 19 : 18}>
-							<h5 style={{ display: 'inline-block' }}>
-								{isEditable ? 'Edit JSON' : 'JSON Result'}
-							</h5>
-						</Col>
-						<Col span={isEditable ? 5 : 6}>
-							<Tooltip visible={copied} title="Copied">
-								<Button
-									shape="circle"
-									icon="copy"
-									size="small"
-									onClick={() => this.copyJSON(res)}
-								/>
-							</Tooltip>
-							{isEditable ? (
-								<Button
-									size="small"
-									type="primary"
-									style={{ marginLeft: '5px' }}
-									disabled={!isValidJSON}
-									onClick={() => this.handleUpdateJSON(editorValue)}
-								>
-									Update
-								</Button>
-							) : (
-								<Button
-									size="small"
-									type="primary"
-									style={{ marginLeft: '5px' }}
-									disabled={!isValidJSON}
-									onClick={() => this.handleEditing()}
-								>
-									Edit
-								</Button>
-							)}
-						</Col>
-					</Row>
-				)}
-			>
-				<Button shape="circle" icon="file-text" style={{ marginRight: '5px' }} />
-			</Popover>
-		);
-	};
+	}
 
 	render() {
 		const {
@@ -512,41 +254,10 @@ export default class Editor extends Component {
 			appName,
 			credentials,
 			url,
-			mappings,
-			customProps,
-			onPropChange,
-			mappingsType,
-			showCodePreview,
-			showCustomList,
 		} = this.props;
 		const {
 			renderKey, showModal, showVideo,
 		} = this.state;
-		let resultComponentProps = componentProps.result || {};
-		resultComponentProps = {
-			size: 5,
-			pagination: true,
-			paginationAt: 'bottom',
-			scrollTarget: 'result',
-			...resultComponentProps,
-			onData: (res, triggerClickAnalytics) => {
-				const { _id, _index, ...renderedJSON } = res;
-				return (
-					<div className={listItem} key={res._id} onClick={triggerClickAnalytics}>
-						<ExpandCollapse previewHeight="390px" expandText="Show more">
-							{<Tree showLine>{this.renderAsTree(renderedJSON)}</Tree>}
-						</ExpandCollapse>
-						<div style={{ marginTop: 10, textAlign: 'right' }}>
-							{this.renderJSONEditor(res)}
-							{this.renderDeleteJSON(res)}
-						</div>
-					</div>
-				);
-			},
-			react: {
-				and: Object.keys(componentProps).filter(item => item !== 'result'),
-			},
-		};
 		const title = (
 			<span>
 				Search Preview{' '}
@@ -585,12 +296,7 @@ export default class Editor extends Component {
 									<RSWrapper
 										id={config}
 										component="MultiList"
-										mappings={mappings}
-										customProps={customProps}
-										showCodePreview={showCodePreview}
-										showCustomList={showCustomList}
 										componentProps={componentProps[config] || {}}
-										onPropChange={onPropChange}
 										onDelete={this.props.deleteComponent}
 										full
 									/>
@@ -604,12 +310,7 @@ export default class Editor extends Component {
 								component={
 									this.props.useCategorySearch ? 'CategorySearch' : 'DataSearch'
 								}
-								mappings={mappings}
-								customProps={customProps}
-								showCodePreview={showCodePreview}
-								showCustomList={showCustomList}
 								componentProps={componentProps.search || {}}
-								onPropChange={onPropChange}
 							/>
 						</Card>
 
@@ -619,15 +320,15 @@ export default class Editor extends Component {
 								id="result"
 								component="ReactiveList"
 								key={renderKey}
-								mappings={mappings}
-								customProps={customProps}
-								mappingsType={mappingsType}
-								showCodePreview={showCodePreview}
-								showCustomList={showCustomList}
-								componentProps={resultComponentProps}
-								renderJSONEditor={this.renderJSONEditor}
-								renderDeleteJSON={this.renderDeleteJSON}
-								onPropChange={onPropChange}
+								componentProps={
+									componentProps.result ? {
+											...componentProps.result,
+											react: {
+												and: Object.keys(componentProps).filter(item => item !== 'result'),
+											},
+										} : {}
+									}
+								setRenderKey={this.setRenderKey}
 								full
 								showDelete={false}
 							/>
